Adds PostgreSQL-backed JSONB snapshot persistence for the financials API while preserving JSON file persistence as the default fallback.

Includes Flyway migration support, postgres profile configuration, pgAdmin-based setup documentation, ADR/roadmap updates, and CI lockfile fixes.

+++ backend/src/main/java/com/example/backend/dto/financials/IncomeSummaryItemResponse.java
@@ -1,3 +1,6 @@
 package com.example.backend.dto.financials;
 
-public record IncomeSummaryItemResponse(long id, String category, String interval, double amount) {}
+import java.math.BigDecimal;
+
+public record IncomeSummaryItemResponse(
+    long id, String category, String interval, BigDecimal amount) {}
