+++ backend/src/main/java/com/example/backend/repository/ExpenseBill.java
@@ -1,7 +1,9 @@
 package com.example.backend.repository;
 
+import java.math.BigDecimal;
+
 public record ExpenseBill(
-    long id, String bill, int dueDay, double amount, String account, boolean paid) {
+    long id, String bill, int dueDay, BigDecimal amount, String account, boolean paid) {
 
   public ExpenseBill withId(long id) {
     return new ExpenseBill(id, bill, dueDay, amount, account, paid);
