+++ backend/src/main/java/com/example/backend/repository/IncomeSummaryItem.java
@@ -1,6 +1,8 @@
 package com.example.backend.repository;
 
-public record IncomeSummaryItem(long id, String category, String interval, double amount) {
+import java.math.BigDecimal;
+
+public record IncomeSummaryItem(long id, String category, String interval, BigDecimal amount) {
 
   public IncomeSummaryItem withId(long id) {
     return new IncomeSummaryItem(id, category, interval, amount);
